# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: KNXVault API
  version: 0.4.7
  description: Secrets management and PKI REST API (Phase 1 MVP + Phase 2 complete).
  license:
    name: Apache 2.0
    identifier: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8200
    description: Local development
...
paths:
...
  /auth/kubernetes:
    post:
      tags: [auth]
      summary: Kubernetes service account login
      operationId: loginKubernetes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/K8sLoginRequest"
      responses:
        "200":
          description: Client token issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
  /pki/issue:
    post:
      tags: [pki]
      summary: Issue a leaf certificate
      operationId: issueCertificate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCertRequest"
      responses:
        "201":
          description: Certificate issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueCertResponse"
...
  /sys/auth/approle:
    post:
      tags: [sys]
      summary: Register AppRole credentials
      operationId: registerAppRole
      security:
        - bearerAuth: []
      responses:
        "200":
          description: AppRole registered
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    K8sLoginRequest:
      type: object
      required: [role, jwt]
      properties:
        role:
          type: string
        jwt:
          type: string
...
    LoginResponse:
      type: object
      properties:
        client_token:
          type: string
        ttl:
          type: integer
        policies:
          type: array
          items:
            type: string
        renewable:
          type: boolean
...
    IssueCertRequest:
      type: object
      required: [role, common_name]
      properties:
        role:
          type: string
        common_name:
          type: string
        dns_names:
          type: array
          items:
            type: string
        ip_addresses:
          type: array
          items:
            type: string
        ttl:
          type: string
        key_bits:
          type: integer
        auto_renew:
          type: boolean
          description: Track and auto-renew this certificate before expiry
...
    IssueCertResponse:
      type: object
      properties:
        cert_pem:
          type: string
        private_key_pem:
          type: string
        serial:
          type: string
        expires_at:
          type: string
...